requirements.txt
colorama==0.4.6
et_xmlfile==2.0.0
exceptiongroup==1.3.0
flake8==7.1.1
iniconfig==2.1.0
mccabe==0.7.0
mypy==1.11.2
mypy_extensions==1.1.0
openpyxl==3.1.2
defusedxml==0.7.1
packaging==25.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pydantic==1.10.8
pyflakes==3.2.0
pytest==8.3.3
python-dotenv==1.0.1
tomli==2.3.0
tqdm==4.66.4
typing_extensions==4.13.2
wheel==0.44.0